general_function.py file added
